rename: zonefile -> bind
