+++ .gitignore
@@ -103,4 +103,5 @@ dist
 # TernJS port file
 .tern-port
 dist/*-grammar.js
+dist/*.js
 package-lock.json
