+++ .npmignore
@@ -11,4 +11,4 @@ codecov.yml
 .eslintrc.yaml
 .eslintrc.json
 package-lock.json
-dist/*-grammar.js
+dist/*.js
