+++ bin/dns-zone
@@ -9,7 +9,7 @@ const cmdLineArgs = require('command-line-args')
 const cmdLineUsage = require('command-line-usage')
 
 const ZONE    = require('../index').ZONE
-const zf      = require('../lib/zonefile')
+const bind    = require('../lib/bind')
 const tinydns = require('../lib/tinydns')
 const maradns = require('../lib/maradns')
 const RR      = require('dns-resource-record')
@@ -20,7 +20,7 @@ const opts = cmdLineArgs(usageOptions())._all
 if (opts.verbose) console.error(opts)
 if (opts.help) usage()
 
-zf.zoneOpts = {
+bind.zoneOpts = {
   origin: rr.fullyQualify(opts.origin) || '',
   ttl   : opts.ttl || 0,
   class : opts.class || 'IN',
@@ -31,7 +31,7 @@ zf.zoneOpts = {
     sameOwner: opts['hide-same-owner'],
   },
 }
-if (opts.verbose) console.error(zf.zoneOpts)
+if (opts.verbose) console.error(bind.zoneOpts)
 
 ingestZoneData()
   .then(r => {
@@ -41,7 +41,7 @@ ingestZoneData()
       case 'maradns':
         return maradns.parseZoneFile(r.data).then(checkZone)
       default:
-        return zf.parseZoneFile(r.data).then(zf.expandShortcuts).then(checkZone)
+        return bind.parseZoneFile(r.data).then(bind.expandShortcuts).then(checkZone)
     }
   })
   .then(output)
@@ -54,7 +54,7 @@ function checkZone (zoneArray) {
   return new Promise((resolve, reject) => {
     try {
       new ZONE({
-        ttl: zf.zoneOpts.ttl, origin: zf.zoneOpts.origin, RR: zoneArray,
+        ttl: bind.zoneOpts.ttl, origin: bind.zoneOpts.origin, RR: zoneArray,
       })
       // console.log(z)
       resolve(zoneArray)
@@ -229,7 +229,7 @@ function ingestZoneData () {
       filePath = process.stdin.fd
     }
     else {
-      if (!opts.origin) zf.zoneOpts.origin = rr.fullyQualify(path.basename(filePath))
+      if (!opts.origin) bind.zoneOpts.origin = rr.fullyQualify(path.basename(filePath))
     }
 
     if (opts.verbose) console.error(`reading file ${filePath}`)
@@ -248,7 +248,7 @@ function output (zoneArray) {
   // console.error(zoneArray)
   switch (opts.export.toLowerCase()) {
     case 'json'   : return toJSON(zoneArray)
-    case 'bind'   : return toBind(zoneArray, zf.zoneOpts.origin)
+    case 'bind'   : return toBind(zoneArray, bind.zoneOpts.origin)
     case 'tinydns': return toTinydns(zoneArray)
     case 'maradns': return toMaraDNS(zoneArray)
     default:
@@ -266,8 +266,8 @@ function isBlank (rr) {
 function toBind (zoneArray, origin) {
   for (const rr of zoneArray) {
     if (isBlank(rr)) continue
-    process.stdout.write(rr.toBind(zf.zoneOpts))
-    zf.zoneOpts.previousOwner = rr.get('owner')
+    process.stdout.write(rr.toBind(bind.zoneOpts))
+    bind.zoneOpts.previousOwner = rr.get('owner')
   }
 }
 
@@ -300,14 +300,14 @@ function toHuman (zoneArray) {
 
   // console.log(widest)
   let rdataWidth = process.stdout.columns - widest.owner - widest.type - 10
-  if (!zf.zoneOpts.hide.ttl) rdataWidth -= widest.ttl
+  if (!bind.zoneOpts.hide.ttl) rdataWidth -= widest.ttl
 
   for (const r of zoneArray) {
     if (isBlank(r)) continue
 
     process.stdout.write(r.get('owner').padEnd(widest.owner + 2, ' '))
 
-    if (!zf.zoneOpts.hide.ttl) {
+    if (!bind.zoneOpts.hide.ttl) {
       process.stdout.write(r.get('ttl').toString().padStart(widest.ttl, ' ') + '  ')
     }
