+++ lib/maradns.js
@@ -12,7 +12,7 @@ exports.parseZoneFile = async str => {
 
   try {
     const nearley = require('nearley')
-    const grammar = nearley.Grammar.fromCompiled(require('../dist/csv2-grammar.js'))
+    const grammar = nearley.Grammar.fromCompiled(require('../dist/mara.js'))
     grammar.start = 'main'
     parser = new nearley.Parser(grammar)
     parser.feed(str)
